+++ ricaun.Nuke.PackageBuilder/Components/IHazOutput.cs
@@ -1,9 +1,7 @@
-﻿using System;
-using Nuke.Common;
+﻿using Nuke.Common;
 using Nuke.Common.IO;
 using Nuke.Common.ProjectModel;
 using Nuke.Common.ValueInjection;
-using ricaun.Nuke.Extensions;
 namespace ricaun.Nuke.Components
 {
     /// <summary>
@@ -18,15 +16,9 @@ namespace ricaun.Nuke.Components
         string Folder => ValueInjectionUtility.TryGetValue(() => Folder) ?? "Output";
 
         /// <summary>
-        /// Project (default: <seealso cref="IHazPackageBuilderProject.GetPackageBuilderProject"/>)
-        /// </summary>
-        [Parameter]
-        public Project Project => ValueInjectionUtility.TryGetValue(() => Project) ?? GetPackageBuilderProject();
-
-        /// <summary>
         /// OutputDirectory
         /// </summary>
-        AbsolutePath OutputDirectory => GetOutputDirectory(Project);
+        AbsolutePath OutputDirectory => GetOutputDirectory(GetPackageBuilderProject());
 
         /// <summary>
         /// GetOutputDirectory